♻️ refactor: 요청&응답 dto 구조 변경
