+++ src/main/java/com/zip/community/platform/adapter/in/web/dto/request/chat/MessageDeleteRequest.java
@@ -1,0 +1,12 @@
+package com.zip.community.platform.adapter.in.web.dto.request.chat;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class MessageDeleteRequest {
+
+    private String messageId;
+    private Long memberId;
+}
