+++ src/main/java/com/zip/community/platform/adapter/in/web/dto/request/chat/MessageReportRequest.java
@@ -1,0 +1,13 @@
+package com.zip.community.platform.adapter.in.web.dto.request.chat;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class MessageReportRequest {
+
+    private String messageId;
+    private Long reportMemberId;
+    private String reason;
+}
